lesson5/requirements.txt
# Voice Agents
livekit>=0.13.0
livekit-agents>=0.8.0
livekit-plugins-openai>=0.10.0  
livekit-plugins-silero>=0.7.0
openai>=1.40.0

# Context Engineering
anthropic>=0.39.0
tiktoken>=0.7.0

# Agent Orchestration
# Note: Tasks system is built into Claude Code CLI
# These are for demo/testing purposes
pydantic>=2.0.0
rich>=13.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
asyncio-throttle>=1.0.0

# Optional: Observability
# lmnr[claude-agent-sdk]>=0.5.0
